refactor: rename OnboardingModal to Onboarding
